Bag knows its own storage paths
In BagMoveJob, get the bag's source (upload) path and destination
(storage) path directly from the bag rather than passing as
parameters to BagMoveJob.

Also, consistently name these things "src_path" and "dest_paths" in
models, jobs, and tests. "upload path" and "storage path" should refer
to the base directories in the config only. Bag#storage_location should
refer to where an ingested bag actually is, not where it should go.

An alternative to this would be to instead continue passing the source &
destination paths to the BagMoveJob and instead have the validation
command use the source path from the BagMoveJob. That would imply bags
shouldn't know what their upload & storage paths are which seems counter
ot the intention of merging Bag and Request.