Views for audits
Includes an expansion option for show, which renders each events for
successes & failures. Doesn't expand packages (although it would be
trivial to do so). Expansion is optional for show since performance may
be slow if there are a large number of items. We may decide we want to
handle "expand=failures" to only render the failures, but right now the
option is boolean only.