Audits:
- Audit mailer & specs for fixity check failures. Currently this is one
per failure. Not sure if we should consider consolidating them; that'd
be a little tricky with the job approach we're taking now, but we could
provide a daily summary of failures or something. We'd expect failures
to be extremely rare unless there was some kind of systemic problem,
which could result in a mail bomb.

- Render eventss in expanded audit with info about the package they
pertain to and without

- Adds audit info to the jobs for pass through to recording events

- Additional specs related to issues found during integration testing -
e.g. adding the location header when creating an audit and setting the from
address