move specs for bad or missing auth token
de-duplicates them from each controller spec and shared example group to
application_controller_spec, which makes mores senss since the
authentication logic is there. That makes changes to the authentication
logic easier to test (i.e. planning for REMOTE_USER) easier to test
since we only need to do it in one place instead of many.

We will probably also want to change the examples in the controllers
from "as an admin user" and "as an underprivileged user" to "as a user
that has permission to do the action" and "as a user that does not have
permission to do the action" as we move to more fine-grained permissions
based on policies.