Add Checkpoint & Keycard-alike support for package show policy
This is just a first step towards policies with the package policy
checking whether the show action has been granted.

The next step is role-based authentication - recording admins in the
permit database rather than in the User role, and a creation role that
allows individual users to upload packages of a particular type. This
would result in all the policies to query Checkpoint

The "UserAttributes" is a temporary construct until Keycard provides
support for getting the username from the request.

The AgentResolver and specs as well as the checkpoint helpers in the
specs are copied directly from swapmeet.