quick hack to add file listing to package show view
This makes the view spec a little bit confusing since the index and show
views now don't behave exactly the same way. (There's kind of a similar
issue with testing the audit view). Adding a package presenter here
could help isolate the conditionality and also remove the dependency of
the view specs on Services.storage.