PFDR-99: Test policies are checked in controller actions
Adds a couple of helpers that mock the policy. It would be nicer to be
able to inject a policy, but controller specs make this hard.
Longer-term, we could think about adding something similar to Pundit
where the controller raises an error if the policy wasn't checked, but
this meets the need for now. This doesn't add anything with respect to
ensuring the scope was properly resolved, but there are existing
controller specs that cover that, albeit using the real policy rather
than a mock one.