Revert "use url for checkpoint/keycard database"
This reverts commit 7f0fecc4a110b1a0cfd6a834c9cb9c08ff389833.

Keycard's and Checkpoint's database connection methods accept both a url
or a hash of connection parameters. This flexibility lets us always just
set the db configuration option to the value from our configs, rather
than explicitly specifying a url.

We are reverting this change because we can still use a url for the
database connection string without it.