PFDR-98 - Restructure index endpoint specs
The shared examples for "an index endpoint" were pretty complicated
because they relied on setting up specific lets and some tricky shared
namespacing between the spec file and the shared examples.

By being able to inject a mock policy, we can test the generic behavior
without any of the complicated setup. This makes the shared example much
simpler. The only information it uses is the name of the policy so the
doubles created as verifying -- so errors are thrown if the mocks and
the real policy classes get out of sync.

The FakePolicy is renamed to PolicyDouble to reflect that it is just an
RSpec double, rather than a more involved fake collaborator. The
ResourceController pattern is also cleaned up with a metaprogramming
shorthand to declare the policy classes rather than requiring method
declarations and assignments in every controller.