PFDR-98 - Clean up "show endpoint" shared examples
This includes a tentative model for shared examples like this. I don't
think we gain enough in this case, but want a commit here for reference.
The shared example block takes parameters to make things explicit, but
they are somewhat complicated by needing to pass additional query
parameters to the GET call. Because the shared example fabricates the
resource, this ends up being a lambda... too complicated for what it
gets us, here.