Use a repository as Services.storage
This was previously just the ChipmunkBag class, and all callers bound
to the #new method on that class. With this change, this is now a
BagRepository, and all callers bind to #create with the same signature.

This change is intended to remove some bindings on ChipmunkBag, so that
we can more readily improve it and Package's functionality.