De-DRY package view specs
The use of the shared examples defined in a_package_view.rb introduced
coupling in the specs that made them harder to understand and to change.
This commit removes the shared examples, and rewrites the specs to
simply include the similar code. This allows the two specs to be changed
independently.

This allowed the removal of PackageFileGetter as a dependency from the
package views.