Enable rubocop as a travis build target
This involves changes to the rubocop config in .rubocop.yml, based on
the rubocop configuration in (https://github.com/mlibrary/project_template).
Fixes performed are minimal, and most issues are not addressed, but are
instead shunted into .rubocop_todo.yml. This is rubocop's recommend
way of handling legacy projects.

This fixes:
* Issues that rubocop could fix automatically
* Issues that rubocop could not automatically ignore via the
  .rubocop_todo.yml generated by the --auto-gen-config flag.

Additionally:
* Enable rubocop as a travis build
* Add rubocop binstub
* Add rubocop-performance