Move job-queuing utilities with jobs; remove autoload
The QueueItemBuilder and RequestBuilder are really about working with
the model objects and queuing jobs, so this colocates them with the
jobs. It may be that the nuts and bolts of the jobs should come to lib
and these builders actually become the ActiveJob implementations, but
that is for later discussion.

With this change, everything under lib is explicitly required by way of
requiring chipmunk.rb in services.rb, so the autoload is removed.