Begin building Packge Repository
This is a rather extensive commit that introduces a package repository
and storage concepts. They are not at all done, but appear to be a good
start. There are TODO comments added with specific thoughts and
challenges emerging.

The design theory here is that there is a repository of packages, and
those packages may be of different types. For each type, there is a
storage adapter registered to map the general hierarchical structure and
operations on that type of package to the specifics of disk, memory, or
object storage specifics. The goal of this approach is to provide a wall
behind which File, Dir, and so on, are not visible, but there are
methods that give strings, hashes, and IO objects/blocks.

The Package interface is yet to be properly extracted, somewhat
challenged by identifier-storage semantics and the extent to which Bag
leaks Pathname objects for use with File operations elsewhere.

The BagRepository is left intact for now because of similar challenges:
it is used in various places to instantiate Bags for items that we would
not consider "in" the repository. These transient or factory-type uses
should not be too difficult to separate from those where we are
retrieving stored objects, but there are still open questions about how
to handle the workflow of inbound or in-migration objects.