Prune BagRepository
This unrolls the indirection that was in place with the BagRepository
for a very temporary class method: Bag::__from_package__. The idea here
is to eliminate the two conflicting repository concepts and focus on
removing File operations from Bag. Once those are done (taking only
relative paths to files in a package and returning IO streams), the
Chipmunk::Package interface can be better defined and inform the
Repository/Storage semantics. The biggest challenge is that the Package
model is part process object, part repository. The problem is not really
that they are AR models, but the business rules of disk access and
lifecycle of deposits, requests, etc., are currently tightly bound.
Untangling that will be essential to a clear design that enables
polymorphic storage and access as distinct from process management.