PFDR-97: Collection policies use checkpoint & roles
Resolves resource collections for collection policies by looking at what
objects the user has been granted permission to show. Standardizes and
clarifies new vs. create.

- update checkpoint to 1.1.1 to get some helpers

- define additional scopes on Package (possibly to be migrated to
something more generic later)

- create? is a Resource-level operation; it checks whether a user has
permission to create a particular resource that has been built but not
yet persisted

- new? is a Collection-level operation; it checks whether a user could
potentially create _some_ new object in the collection, without any
information about the specific object.

- the QueueItemBuilder needs some attention (PFDR-169); the change to
QueueItemsController here is a bit of a messy change so that we can
authorize create? given the built-but-not-persisted object

- some starts at delegation from one policy to another, specifically for
things related to packages (might end up just using the delegated
policies directly, if they have no differences)

- Added some helpers to wipe the database after specs that create
grants. This might be better done by wrapping stuff in a transaction.

- Move resolvers into lib

- PFDR-162: Package.of_any_type - expands to all content types of package for
matching when authorizing a collection-level operation

- rethink CollectionPolicy.can? - now it takes a resource, and there is
a convenience method in CollectionPolicy for
Checkpoint::Resource::AllOfType. We could also perhaps try to
dynamically figure out what we should do with what's passed in similar
to the resource resolver itself - this might be pointing at a notion of
a collection-level resource resolver?