PFDR-169: inline RequestBuilder to PackagesController
Inlines RequestBuilder in a similar fashion to QueueItemBuilder.

Also, authorizes a :show? before issuing a redirect to minimize
information leakage (i.e. an unauthorized user attempting a CREATE on a
thing they don't have permission to show receiving a different status
based on whether the thing exists or not). I'm not sure this is correct
or how much information leakage there should be even for viewers; it
might be better to authorize new? before even looking for a duplicate
item.

Also starts to address some of the vestigial "Request" language, although
now it seems likely we'll start calling Packages "Descriptors" instead,
at least internally