PFDR-170 - Fix policy helpers enough for EventsPolicy
Injects package policy to events & queue items policy

This cleans up a number of issues with the policy helpers so that
it_allows, it_forbids (identical to it_disallows), it_resolves, and
it_has_base_scope helper methods use the allow_action, forbid_action,
resolve, and have_base_scope matchers.

There was an attempted convenience in being able to use the
described_class as the subject, but this caused the matchers to be more
coupled and complex than needed. It isn't that much work to create
policy instances as the subject, and this is much more aligned with
general practice.

This is enough to get the EventsPolicy cleaned up and delegating to the
PackagesPolicy with tidy tests. The others should require only mostly
mechanical changes to get their tests running again.