PFDR-170 - Finish fixing up policy delegation
 - Add ViewableResources
 - Ensure that EventsPolicy and QueueItemsPolicy delegate to
   PackagesPolicy properly
 - Fix various bugs and inconsistencies in the matchers/helpers

The biggest change here is that the CollectionPolicy no longer has the
scope_for_resources stuff, extracting that to the ViewableResources
class, as a sort of filter for relations. That moves the resource_types
responsibility back to the relation, rather than on the concrete policy.

This ended up being a bigger commit than anticipated because the matcher
and FakeCollection needed some work to be stable with the chained OR
scopes, and that forced a few downstream changes.