PFDR-182 - Support multiple formats in BagRepository/PackageStorage
This commit represents the majority of the format abstraction for the
app. Everything else was already using BagRepository through the
`storage` service, so this will catch all instantiations of
Chipmunk::Bag outside of any specs that use it directly. It was not
really a repository, so I have renamed it to be aligned better with its
responsibility as a storage proxy factory.

 - Rename BagRepository to PackageStorage
 - Change initializer parameter to a hash of formats to storage classes
 - Restructure test to have one scenario and use behavioral style
 - Add PackageNotStoredError