PFDR-183 - Make audit job format-indifferent
Remove all references to "bag" in the AuditFixityCheckJob. This implies
the validation contract of any storage proxy:

 - #valid? validates the object
 - #errors holds a collection of validation errors
 - that #errors has a #full_messages list that can be .join'd

We should consider extracting a validation interface. It's convenient to
have the method directly on the object, but it makes it stateful in an
odd way, setting up a side effect between the valid? and errors method.
An immutable validation result is almost certainly better, whether or
not it is generated internally or from an external validator.