PFDR-184 - Separate BagMoveJob from Bag::Validator
This is an incomplete transition toward separating the ingest and
storage phases for validation purposes. It cleans up BagMoveJob
considerably, though. It seems likely that this job will transition to
a generic ingest role, validating the inbound package (in whatever
format) and then ingesting it (to whatever format) with other
components based on the deposit.

The src_path on Package is still difficult to deal with in that it is
bound to the stored state and upload path conventions of including the
username. This results in some nasty mocking and stubbing.

 - Add (assumed) temporary valid_for_ingest? method to Package
 - Clean up BagMoveJob to take only the QueueItem again
 - Restructure BagMoveJob so the flow is much clearer and receptive to
   further extraction of validation or moving based on format