PFDR-185 - Add storage_volume and related handling
This adds the notions of a named Volume and a VolumeManager component to
serve as a registry. I've used an approach where there is a "root"
volume to help make the transition from Packages holding their full
storage_location to using abbreviated paths within a symbolically named
volume. For now, we do not need to decide the destination volume or
modify the storage_location -- we use "root" and use the absolute path
as pseudo-relative.

We simply set the volume to root for any existing packages upon database
migration and any successful runtime ingests. Later, when we extract the
decision of destination volume and the mechanics of moving/copying
between volumes, we can configure another one and execute a runtime
migration from the root volume to it (thereby localizing the paths),
after which the root volume can be removed.