PFDR-185 - Fix up specs related to storage_volume
There were a few places in the tests relying on the knowledge that
storage_location was the condition for stored or that the base package
fabricator had a storage_location. I can go either way on whether the
assumption is that the base fabricator gives a stored or unstored
package, but it was inconsistent. This also tests that the "root" volume
is set when moving a bag (a transitional item; see the previous commit).