PFDR-189 - Consider storage_volume in stored? calculation
Also resolves PFDR-191.

This commit nearly finishes the abstraction from the absolute paths for
packages.

 - Package.stored scope requires non-null volume and location
 - Package#stored? requires non-null volume and location
 - Package#storage_path returns the storage_location attribute value
 - Package#storage_location returns the full path (volume + path)
 - Package#storage_location is expected to be written with a relative
   path; while the root volume is used, these equal absolute

We will need to audit the places where the storage_location is used as a
full path from the package and decide whether or not that should
continue. I suspect that we will encapsulate the IO through
PackageStorage, looking up the the appropriate volume there, and then
all of the actual root and disk path details will be outside of Package.