Implement scenario "Uploading an artifact"
This involved setting before and after hooks in
features/support/sandbox.rb to create the incoming storage directory,
and to clean it up afterwards.

We also add config in features/support/env.rb to configure the ActiveJob
queue adapter to be the test adapter. This forces ActiveJobs to run
in-line; they do not rely on a backend adapter. (Without this, we'd need
something like Redis running.)