Use noop external validators in unit tests
As these require a working feed installation, we only want to run
them when RUN_INTEGRATION is set to 1. This hasn't been a problem so
far, but that's only because we've avoided it or mocked it out
individually. Doing it this way is consistent, and enables things like
acceptance tests to run without requiring the external validators.