PFDR-208 - Clean up login handling
This commit restructures how logins/sessions are handled pretty
significantly. It uses the new constructs in Keycard and trims back the
provisional handling that was built-in here (like setting @current_user
and adding to the user identity).

There is a development-mode login form that allows becoming any user,
with a configuration setting, allow_impersonation.

The UsersController is removed, replaced by a LoginController, with more
consistent handling of when requests are considered authenticated and
when/where they should be redirected.