PFDR-237 - Add versioning features not tied to the API
This introduces some features at a higher level than the v2-specific
ones. I've left the more detailed ones in place, but I think those
should probably be translated in to request specs or maybe moved to the
chipmunk-client. It doesn't hurt to have them as features, but they are
more protocol-oriented, where the main advantage of this format is to
abstract from technical details and present user-facing details. In this
case, deposits and revisions are more coarse-grained transactions than
exposed at the API (for technical reasons of size and transfer only).
The features at the top-level directory attempt to stay focused on the
details that are important to the people who would have the named roles.