Use reader and writer objects in Volume
* Introduces the concept of a Reader, used to load instances from disk.
  This immediately replaces Volume's call to Bag.new with delegation to
  the volume's reader, an instance of Bag::Reader

  This conceptoffers a point from which we can more effectively handle
  different types of stored objects.

* Similarly, we introduce the concept of a Writer, implement
  Bag::MoveWriter, and use it in our volumes.

The main driver here is to have the ability to read and write v1 bags,
versioned bags, and in the future OCFL objects in a consistent way. This
allows us to use PackageStorage for all of these object
types, avoiding having something like RevisionStorage and OCFLStorage.

This approach was chosen because the responsibility for simply
writing the object felt like it belonged in Volume, and so we inject it
as a dependency. Meanwhile, the high-level rules around writing,
especially where it fits in the lifecycle, belong to PackageStorage.