Use Bag::MoveWriter to write instances to disk
The main driver here is to have the ability to read and write v1 bags,
versioned bags, and in the future OCFL objects in a consistent way. The
immediate result is I hope to use PackageStorage for all of these object
types, avoiding having something like RevisionStorage and OCFLStorage.

This approach was chosen because it the responsibility for simply
writing the object felt like it belonged in Volume, and so we inject it
as a dependency. Meanwhile, the high-level rules around writing,
especially where it fits in the lifecycle, belong to PackageStorage.