Prefer storage_format over format
Rails overwrites "format" in the request's parameter hash based on the
body of the request, e.g. json. We didn't notice this before because
we set the format of the created Package to Chipmunk::Bag.format.
Now that we want to get the format from the user, we run into this
conflict.

This works around the issue by renaming the field to storage_format

It also removes the paranoid format check from Volume. That sort of
check is now part of validation.