Force Travis-CI to actually run migrations
Previously, this relied on Rails's automatic behavior to migrate the
database, which seems to build the database from schema.rb. As schema.rb
is not created correctly when the PK of a table is a uuid, it would skip
creating the artifacts table.

By forcing the migrations to run, we better test the production setup,
and avoid this issue.