PFDR-277 - Add rake task to restart lost jobs
We've seen Resque jobs crash occasionally, which leaves QueueItems in
pending status but with nothing in the queue, so they are permanently
orphaned. It's apparently because of some strange threading deadlock
with semantic_logger. Those we have enqueued again have finished.

This rake task is intentionally naive. It checks for pending jobs and
the queue. If there are no jobs running but any are marked pending, they
are considered stranded and the oldest is put back into the queue.