PFDR-279 - Add queue item status to API
This change introduces a new route, /v1/queue/status.json, that gives
status of all QueueItems for Packages someone may view. It adds the
'jb' gem because jbuilder with partials and collections performs very
badly (on the order of 20 seconds for production-scale data). The
endpoint could have pagination, but that is unneeded complication at
this point.